% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{CreateCREscendoObject}
\alias{CreateCREscendoObject}
\title{Create a CREscendo Object}
\usage{
CreateCREscendoObject(
  annotated_peaks,
  cleavage_sites_counts,
  CRE_annotations,
  peaks_annotated_with_CREs,
  multiCREs
)
}
\arguments{
\item{annotated_peaks}{A data frame containing information about the annotated peaks.}

\item{cleavage_sites_counts}{A list of matrix with counts of cleavage sites across different peak regions.}

\item{CRE_annotations}{Data frame detailing annotations of CREs, including type, location, etc.}

\item{peaks_annotated_with_CREs}{A GenomicRange that links peaks to specific CRE annotations.}

\item{multiCREs}{A logical value indicating whether to include peaks with multiple CREs (TRUE) or only peaks with one CRE (FALSE).}
}
\value{
An object of class 'CREscendo' containing the specified datasets, structured for further analysis.
}
\description{
This constructor function initializes a CREscendo object containing various components
related to cis-regulatory element (CRE) analysis. The object consolidates annotated peaks,
counts of cleavage sites, CRE annotations, and peaks annotated with CREs.
}
