% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{CreateCREscendoTestedObject}
\alias{CreateCREscendoTestedObject}
\title{Create a CREscendoTested Object}
\usage{
CreateCREscendoTestedObject(CREscendo, celltypes, annotated_tests_on_peaks)
}
\arguments{
\item{CREscendo}{A CREscendo object containing initial data such as annotated peaks and
cleavage sites counts.}

\item{celltypes}{A vector of cell types for which tests were conducted. This parameter also
automatically includes a negation of each cell type (e.g., "non-Type1" if "Type1" is specified)
if only one cell type is provided.}

\item{annotated_tests_on_peaks}{A list containing results of tests on peaks, including chi-squared
test results and summaries.}
}
\value{
A CREscendoTested object that contains detailed results and annotations from CRE testing
processes, structured for further analysis or visualization.
}
\description{
This constructor function initializes a CREscendoTested object, which encapsulates the results
and relevant data from CRE association tests. This object includes information about cell types,
chi-squared test results, annotated peaks, cleavage sites counts, CRE annotations, and
peaks associated with CREs.
}
