% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crescendo.r
\name{crescendo}
\alias{crescendo}
\title{Implementation of the Crescendo algorithm.}
\usage{
crescendo(
  Ycounts,
  meta_data,
  R,
  batch_var = "batch",
  genes_use = NULL,
  prop = 1,
  min_cells = 50,
  lambda = NULL,
  alpha = 0,
  seed = 2,
  mc.cores = NULL,
  impute_reads = NULL,
  return_obj = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Ycounts}{A matrix containing un-normalized gene count information (cells x genes).}

\item{meta_data}{Metadata table that contains cell information (cell x info).}

\item{R}{A matrix of harmony clusters (cluster x cell) that are received as output from the Harmony algorithm.}

\item{batch_var}{Name of column containing batch information (e.g. batch, dataset, technology).}

\item{genes_use}{Genes to estimate. If none supplied, will automatically estimate all genes in Ycounts.}

\item{prop}{Fraction of cells to downsample to (e.g. 0.5 means downsample to arouond 50\% of original cells, 0.25 means 25\%, etc.).}

\item{min_cells}{Number of minimum cells to keep from each cell type in each batch.}

\item{lambda}{Regularization parameter.}

\item{alpha}{Elastic net regularization parameter. 1 indicates lasso penalty, 0 indicates ridge penalty.}

\item{seed}{Seed for RNG.}

\item{mc.cores}{Number of cores to use for parallel functions.}

\item{impute_reads}{If not NULL, supplies a vector of UMI to perform imputation at the supplied read depths.}

\item{return_obj}{Return only corrected counts, or results from each step.}

\item{verbose}{Print out progress messages.}
}
\value{
A dataframe containing corrected gene count expressions, or a list containing results from each Crescendo step.
}
\description{
Runs the Crescendo algorithm by performing the downsampling (if necessary), estiimation, marginalization, and matching steps.
}
