% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{CREtest.CREscendo}
\alias{CREtest.CREscendo}
\title{Test CRE Association with specified cell types within a CREscendo Object}
\usage{
\method{CREtest}{CREscendo}(x, ..., celltypes = NULL)
}
\arguments{
\item{x}{A CREscendo object containing the necessary datasets for CRE analysis.}

\item{...}{Additional arguments passed to lower-level functions, such as plotting or statistical functions.}

\item{celltypes}{One or two cell types. Must be cell types included in the CREscendo object.
If one cell type is specified, the comparison will be made between the specified cell type and all other cell types.
If two cell types are specified, the comparison will be made between the two specified cell types.}
}
\value{
Returns a CREscendoTestedObject containing the results of the CRE tests,
which include statistical summaries and any relevant annotations for further analysis.
}
\description{
This function performs chi-squared tests on cis-regulatory elements (CREs) associated with peaks
within a CREscendo object. It evaluates differential CRE activities between two cell types.
This is intended to facilitate the analysis of CRE dynamics and their potential regulatory impacts.
}
