% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{CREtest}
\alias{CREtest}
\title{Generic function for CREtest}
\usage{
CREtest(x, ...)
}
\arguments{
\item{x}{An object to be tested.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
This function tests CRE associations in various objects.
}
