% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.r
\name{estimate_betas}
\alias{estimate_betas}
\title{Estimate gene structue with a Poisson generalized linear model.}
\usage{
estimate_betas(
  Ycounts,
  R,
  design_full,
  lambda = NULL,
  alpha = 0,
  genes_use = NULL,
  offset = NULL,
  mc.cores = NULL
)
}
\arguments{
\item{Ycounts}{A matrix containing un-normalized gene count information (cells x genes).}

\item{R}{A matrix containing cluster assignments for each cell (cluster x cell).}

\item{design_full}{A design matrix that contains batch and cluster membership information.}

\item{lambda}{Regularization parameter.}

\item{alpha}{Elastic net regularization parameter. 1 indicates lasso penalty, 0 indicates ridge penalty.}

\item{genes_use}{Genes to estimate. If none supplied, will automatically estimate all genes in Ycounts.}

\item{offset}{A vector containing the total nUMI for each cell.}

\item{mc.cores}{Number of cores to use for parallel functions.}
}
\value{
A dataframe containing parameter estimates for each gene.
}
\description{
Fit a gene's expression distribution with a Poisson generalized linear model. Default is ridge penalty.
}
