% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Extract.CREscendoTested}
\alias{Extract.CREscendoTested}
\title{Extract data of a specific peak or a specific gene from a CREscendoTested object}
\usage{
\method{Extract}{CREscendoTested}(x, ..., which_peak = NULL, which_gene = NULL, plot_contribution = TRUE)
}
\arguments{
\item{x}{A CREscendoTested object containing results from differential CRE tests.}

\item{...}{Additional arguments passed to lower-level functions, such as plotting or statistical functions.}

\item{which_peak}{A character string or numeric identifier specifying the peak for visualization.}

\item{which_gene}{A character string specifying the gene symbol for which visualization is desired.}

\item{plot_contribution}{A logical value indicating whether to plot the contribution of each CRE to the Chi-square statistics.}
}
\value{
A list containing extracted data for further analysis.
}
\description{
This function extracts relevant data from a CREscendoTested object, including cleavage sites, CRE annotation, and CRE frequencies within tested cell types.
}
\details{
The function first checks that the input is a `CREscendoTested` object. If not,
it stops with an error message. It then extracts the data for the specified peak or the peaks within a specific gene from
the object, including chi-squared test results and cleavage sites counts. It then calls internal functions to quantify contribution of each CRE to the Chi-square statistics.
}
