% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Extract}
\alias{Extract}
\title{Generic function for Extract}
\usage{
Extract(x, ...)
}
\arguments{
\item{x}{An object to extract from.}

\item{...}{Additional arguments affecting the extracting.}
}
\description{
This function provides a framework for extracting information of certain peak or gene from objects.
}
