% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Filter.CREscendoTested}
\alias{Filter.CREscendoTested}
\title{Filter the Results of CREscendo Tested Object}
\usage{
\method{Filter}{CREscendoTested}(x, ..., fdr_cutoff = 0.05, coverage_cutoff = 100, abs_diff_cutoff = 0.2)
}
\arguments{
\item{x}{A `CREscendoTested` object containing results from CRE association tests.}

\item{...}{Additional arguments passed to lower-level functions, such as plotting or statistical functions.}

\item{fdr_cutoff}{A numeric value specifying the fdr adjusted p-value threshold for filtering the results. The default is 0.05.}

\item{coverage_cutoff}{A numeric value specifying the cut-off of the number of fragments across all cells within the same type for each peak region for filtering the results. The default is 100.}

\item{abs_diff_cutoff}{A numeric value specifying the cut-off of the absolute difference in the CRE frequencies between two cell types. The default is 0.2.}
}
\value{
Returns a data frame containing
the summary of chi-squared tests conducted on CREs after applying several filters.
The test results are sorted by their chi-squared test statistics in descending order.
}
\description{
This function retrieves the summary of chi-squared test results from a `CREscendoTested` object.
It is specifically used to extract and return high confident results from chi-squared test summary component,
allowing for prioritized downstream analysis.
}
