% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Filter}
\alias{Filter}
\title{Generic function for Filter}
\usage{
Filter(x, ...)
}
\arguments{
\item{x}{An object to filter.}

\item{...}{Additional arguments affecting the filtered results.}
}
\description{
This function provides a framework for filtering various types of objects.
}
