% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalization.r
\name{marginalize_effects}
\alias{marginalize_effects}
\title{Remove effects of confounding factors from a Poisson generalized linear model.}
\usage{
marginalize_effects(
  R,
  design_full,
  betas,
  terms_keep,
  terms_remove_crossed,
  offset
)
}
\arguments{
\item{R}{A matrix containing cluster assignments for each cell (cluster x cell).}

\item{design_full}{A design matrix that contains batch and cluster membership information.}

\item{betas}{A dataframe containing parameter estimates for each gene.}

\item{terms_keep}{A vector containing the terms that we want to keep in the marginalized model.}

\item{terms_remove_crossed}{A vector containing the terms we want to marginalize out (e.g. crossed terms between a confounding
factor and clusters).}

\item{offset}{A vector containing the nUMI to sample up to.}
}
\value{
A dataframe containing rates from the marginalized model for each cell.
}
\description{
Remove effects of confounding factors from a Poisson generalized linear model (obtained from the estimation step) to obtain a 
marginalized model.
}
