% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.r
\name{match_cdfs}
\alias{match_cdfs}
\title{Obtain corrected counts.}
\usage{
match_cdfs(expected_counts_marginalized, Ycounts, design_full, betas, offset)
}
\arguments{
\item{expected_counts_marginalized}{A dataframe containing rates from the marginalized model for each cell.}

\item{Ycounts}{A matrix containing un-normalized gene count information (cells x genes).}

\item{design_full}{A design matrix that contains batch and cluster membership information.}

\item{betas}{A dataframe containing parameter estimates for each gene.}

\item{offset}{A vector containing the nUMI to sample up to.}
}
\value{
A dataframe containing corrected gene expression counts.
}
\description{
Obtain corrected counts by performing inverse cumulative distribution matching.
}
