% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{PrepareCREscendoObject}
\alias{PrepareCREscendoObject}
\title{Prepare a CREscendo Object for Analysis}
\usage{
PrepareCREscendoObject(
  peaks_gr,
  tabix_file,
  cell_annotation,
  celltype_list,
  CRE_gr,
  tssRegion = c(-1000, 1000),
  TxDb,
  annoDb = "org.Hs.eg.db",
  multiCREs = TRUE
)
}
\arguments{
\item{peaks_gr}{GenomicRanges object containing genomic peak locations.}

\item{tabix_file}{Path to a Tabix-indexed file containing fragments relevant to the peaks.}

\item{cell_annotation}{Data frame containing cell annotations with columns for cellname and celltype.}

\item{celltype_list}{Optional; a vector of cell types to include in the analysis. If NULL,
all unique cell types from cell_annotation will be used.}

\item{CRE_gr}{GenomicRanges object for cis-regulatory elements.}

\item{tssRegion}{Numeric vector of length 2 specifying the region around transcription start sites
to consider for annotation (defaults to c(-1000, 1000)).}

\item{TxDb}{A TxDb object containing transcript annotations.}

\item{annoDb}{Character string, the Bioconductor annotation package for organism specific data, default
is "org.Hs.eg.db" for human.}

\item{multiCREs}{Logical value indicating whether to include peaks with multiple CREs (TRUE) or all peaks (FALSE).}
}
\value{
A CREscendo object containing annotated peaks, CRE associations, and cleavage sites counts
        tailored for further analysis.
}
\description{
This function prepares and constructs a CREscendo object by processing genomic peaks,
associating them with cis-regulatory elements (CREs), annotating these peaks,
and compiling counts of cleavage sites for specified cell types. It integrates multiple
genomic and annotation resources to create a comprehensive data object for further analysis.
}
