% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.CREscendo}
\alias{print.CREscendo}
\title{Print method for CREscendo objects}
\usage{
\method{print}{CREscendo}(x, ...)
}
\arguments{
\item{x}{A CREscendo object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function provides a concise summary of a CREscendo object when it is printed to the console.
It displays the number of peaks that overlap with more than one cis-regulatory element (CRE)
and lists the cell types for which fragment cleavage sites have been extracted.
}
\details{
The output includes:
- The number of peaks that have overlaps with multiple CREs.
- A list of cell types for which fragment cleavage site data is available.
}
