% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.r
\name{sample_harmony_clus}
\alias{sample_harmony_clus}
\title{Sample a discrete harmony cluster based off of soft-cluster probabilities.}
\usage{
sample_harmony_clus(R, seed = NULL, mc.cores = NULL)
}
\arguments{
\item{R}{A matrix of harmony clusters (cluster x cell) that are received as output from the Harmony algorithm.}

\item{seed}{Seed for RNG.}

\item{mc.cores}{Number of cores to use for parallel functions.}
}
\value{
A vector containing a discrete cluster assignment for each cell design matrix that contains batch and cluster membership 
information.
}
\description{
Harmony soft-clusters give probabilities for each cell. In order to downsample data, we sample a discrete cluster for each cell
based on its harmony cluster probabilities.
}
