% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Summarize.CREscendoTested}
\alias{Summarize.CREscendoTested}
\title{Summarize the Results of CREscendo Tested Object}
\usage{
\method{Summarize}{CREscendoTested}(x, ...)
}
\arguments{
\item{x}{A `CREscendoTested` object containing results from CRE association tests.}

\item{...}{Additional arguments passed to lower-level functions, such as plotting or statistical functions.}
}
\value{
Returns a data frame containing
the summary of chi-squared tests conducted on CREs.
}
\description{
This function retrieves the summary of chi-squared test results from a `CREscendoTested` object.
It is specifically used to extract and return the chi-squared test summary component,
allowing for quick access to the statistical analysis results.
}
