% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Summarize}
\alias{Summarize}
\title{Generic function for Summarize}
\usage{
Summarize(x, ...)
}
\arguments{
\item{x}{An object to summarize.}

\item{...}{Additional arguments affecting the summary.}
}
\description{
This function provides a framework for summarizing various types of objects.
}
