% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{Visualize}
\alias{Visualize}
\title{Generic function for Visualize}
\usage{
Visualize(x, ...)
}
\arguments{
\item{x}{An object to visualize.}

\item{...}{Additional arguments affecting the visualization.}
}
\description{
This function provides a framework for visualizing various types of objects.
}
