% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.r
\name{downsample_batch_clus}
\alias{downsample_batch_clus}
\title{Downsample data in a batch and cell-type aware manner.}
\usage{
downsample_batch_clus(
  meta_data,
  id_col = "CellID",
  batch_col = "batch",
  clus_col = "sampled_clus",
  prop,
  min_cells,
  seed = NULL
)
}
\arguments{
\item{meta_data}{Metadata table that contains cell information (cell x info).}

\item{id_col}{Name of column containing unique cell IDs.}

\item{batch_col}{Name of column containing batch information (e.g. batch, dataset, technology).}

\item{clus_col}{Name of column containing cluster information.}

\item{prop}{Fraction of cells to downsample to (e.g. 0.5 means downsample to arouond 50\% of original cells, 0.25 means 25\%, etc.).}

\item{min_cells}{Number of minimum cells to keep from each cell type in each batch.}

\item{seed}{Seed for RNG.}
}
\value{
A metadata table containing the downsampled cells.
}
\description{
To downsample data, we ensure that each cell-type in each batch is represented by a minimum number of cells (or if lower than
the minimum, all of the cells).
}
